requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap-form==3.4
django-taggit==1.3.0
gunicorn==20.0.4
Pillow==8.0.1
psycopg2-binary==2.8.6
python-slugify==4.0.1
pytz==2020.1
sqlparse==0.4.1
text-unidecode==1.3
whitenoise==5.2.0
